+++ render-demo/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Updated dependencies
   - @byrding/core@0.7.0
-  - @byrding/react@0.2.4
+  - @byrding/react@0.3.0
 
 ## 0.0.7